ex 4.16-4.21 done

+++ part4/blogList/controllers/blogs.js
@@ -19,19 +19,12 @@ blogsRouter.get('/:id', async (request, response, next) => {
 	}
 });
 
-const getTokenFrom = (request) => {
-	const authorization = request.get('authorization');
-	if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-		return authorization.substring(7);
-	}
-	return null;
-};
-
 blogsRouter.post('/', async (request, response, next) => {
 	const body = request.body;
-	const token = getTokenFrom(request);
-	const decodedToken = jwt.verify(token, process.env.SECRET);
-	if (!token || !decodedToken.id) {
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	console.log('DECODED', decodedToken);
+	console.log('TOKEN', request.token);
+	if (!request.token || !decodedToken.id) {
 		return response.status(401).json({ error: 'token missing or invalid' });
 	}
 	const user = await User.findById(body.userId);
@@ -56,8 +49,17 @@ blogsRouter.post('/', async (request, response, next) => {
 });
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-	await Blog.findByIdAndRemove(request.params.id);
-	response.status(204).end();
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (!request.token || !decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' });
+	}
+	const blog = await Blog.findById(request.params.id);
+	const user = await User.findById(decodedToken.id);
+	if (blog.user.toString() === user.id.toString()) {
+		await Blog.findByIdAndRemove(request.params.id);
+		response.status(204).end();
+	}
+	return response.status(401).json({ error: 'authorization failed' }).end();
 });
 
 blogsRouter.put('/:id', async (request, response) => {
